dynamicimageapp/requirements.txt
click==7.1.1
cycler==0.10.0
dnspython==1.16.0
Flask==1.1.1
Flask-PyMongo==2.3.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.0
numpy==1.18.1
pandas==1.0.1
pymongo==3.10.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
Werkzeug==1.0.0
